layout/inspector/tests/test_isinheritableproperty.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=699592
-->
<head>
  <title>Test for InspectorUtils::isInheritedProperty</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<pre id="test">
<script type="application/javascript">

const InspectorUtils = SpecialPowers.InspectorUtils;

function do_test() {
  is(InspectorUtils.isInheritedProperty("font-size"), true, "font-size is inherited.");

  is(InspectorUtils.isInheritedProperty("min-width"), false, "min-width is not inherited.");

  is(InspectorUtils.isInheritedProperty("font"), true, "shorthand font property is inherited.");

  is(InspectorUtils.isInheritedProperty("border"), false, "shorthand border property not inherited.");
  is(InspectorUtils.isInheritedProperty("garbage"), false, "Unknown property isn't inherited.");

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(do_test);


</script>
</pre>
</body>
</html>
